backend/requirements.txt
fastapi==0.111.0
mangum==0.17.0
httpx==0.27.0
beautifulsoup4==4.12.3
openai==1.35.7
boto3==1.34.144
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
uvicorn==0.30.1
pydantic==2.8.2
